requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.8.0
APScheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
blessed==1.20.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
distro==1.9.0
Django==5.1.6
django-allauth==65.4.1
django-apscheduler==0.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-paypal==2.1
django-picklefield==3.2
django-storages==1.14.4
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docstring_parser==0.16
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.8.2
joblib==1.4.2
kombu==5.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
openai==0.28.0
packaging==24.2
paypalrestsdk==1.13.3
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rsa==4.9
safetensors==0.5.3

service-identity==24.2.0
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.18.3
zope.interface==7.2
